Start to move WM CTS setUp to TestRule
The setUp method being @Before makes it difficult to use
other test rules in conjunction with WindowManager tests,
because the setUp of ActivityManagerTestBase performs some
paranoid checks such as removing all non-launcher tasks
from the screen.

In this patch I'm enabling the TestRule only for one class.
A follow-up should switch the default to this for every
subclass.

Flag: EXEMPT TEST_ONLY
Bug: 383590347
Test: atest BlurTests        # with conversion
Test: atest DisplaySizeTest  # without conversion
Change-Id: I0f4fd21b95eb526b6e2c8b043964aacf4cef4aca