Move AM/WM CTS setUp fully to TestRule
The setUp method being @Before makes it difficult to use
other test rules in conjunction with WindowManager tests,
because the setUp of ActivityManagerTestBase performs some
paranoid checks such as removing all non-launcher tasks
from the screen.

This is a followup after the TestRule only for one class.
Now we switch to this for every subclass.

Flag: EXEMPT TEST_ONLY
Bug: 383590347
Test: atest BlurTests
Test: atest DisplaySizeTest
Change-Id: I35c09e0395b768ff368d1f7b54388bf4d476da24