Fix ClipboardManagerTest for visible background users.
UiDevice#pressHome() does not work well for visible background users, so
explicitly send the HOME event as the visible background user using the
startActivityAsUser() method intead of sending the HOME key event.

This does not hurt what the tests are supposed to test, and this change
only affects devices with visible background user enabled.

Bug: 457904170
Test: atest CtsContentTestCases:android.content.cts.ClipboardManagerTest
      atest --user-type secondary_user_on_secondary_display \
        CtsContentTestCases:android.content.cts.ClipboardManagerTest

Change-Id: Ida7fcbddd3e24da6c100bb09e62e733428b0edca