Require GPU data source queries and trace commands to succeed.
This is a simpler and more robust approach. Instead of repeatedly trying to query data sources, we first wait for activity to actually start and then check the sources are present.
We then do not expect these queries to fail, and it's a helpful signal if they do as profiling data should not be flaky.
See https://googleplex-android-review.git.corp.google.com/c/platform/cts/+/36952147 for the original discussion.

Also adds a check that the activity start succeeded.

Change-Id: I271b0ecd87eb49a8ee2bd707a20d7333c14e52c7
Test: atest -c CtsGpuProfilingDataTest
Flag: EXEMPT TEST_ONLY
Bug: 448868135