Fix ClipboardManagerTest for visible background users.
(This is the second attempt that rolls forward commit
1c0615aba9babdf726a4c3eae6a2d55bf79aa304 after fixing the failrues with
Ravenwood tests. Verified by running CtsContentTestCasesRavenwood.)

UiDevice#pressHome() does not work well for visible background users, so
explicitly send the HOME event as the visible background user using the
startActivityAsUser() method intead of sending the HOME key event.

This does not hurt what the tests are supposed to test, and this change
only affects devices with visible background user enabled.

Bug: 457904170
Test: atest CtsContentTestCasesRavenwood:android.content.cts.ClipboardManagerTest
      atest CtsContentTestCases:android.content.cts.ClipboardManagerTest
      atest --user-type secondary_user_on_secondary_display \
        CtsContentTestCases:android.content.cts.ClipboardManagerTest
Change-Id: I3fd8a09cdcb03417ce9fabaf1054e4200eb00365