Rename @RequireMultiUserSupport
Rename @RequireMultiUserSupport to @RequireSwitchableUserSupport.

The notion of requiring "multiuser" is ambiguous. In this particular
context, for unfortunate historical reasons, it does not refer to
supporting "multiuser" in general (which doesn't even really make sense)
but rather, supporting switchable users. This ties to
UserManager.supportsMultipleUsers(), whose name is stuck as a public
API, but which actually refers to switchable users (as documented
there).

We therefore rename the Bedstead annotation to avoid confusion.

Test: compiles (just a rename)
Bug: 446242633
Flag: EXEMPT BUGFIX
Change-Id: Ib4dc7fb922600f25aa2dee9df2a6be9f794a710e