Correct the Vendor API Level Calculation for NetworkSelectionModeTest
Update the calculation of the NetworkSelectionModeTest to use the
newer date-based format that kicked in during 2024. Without this patch
a device that launched with the 2024 HAL would be incorrectly held
to supporting these CTS tests. While they are only fixing coverage
gaps in the existing, documented API, we do not wish to try and
force vendors to go back and correct implementations that are already
in the field. Hence, only enforcing these tests on devices launcing
with HALs released subsequent to the test changes OR which are not
utilizing the frozen-HAL model.

Bug: 457918984
Test: atest NetworkSelectionModeTest
Flag: EXEMPT BUGFIX
Change-Id: I5c99ea509b9351127359b4788a410931808389c3