cts: usb: add interactive tests to cover Bc12Type and PowerProfileInfo
CtsUsbTypecTestCases adds functional tests to verify whether or not a
device that supports BC 1.2 reporting is capable of detecting an SDP,
CDP, and DCP device as defined in the BC 1.2 specification.

A functional test is also added for devices that support reporting
PowerProfileInfo. testPowerProfileSourceFixed tests whether or not the
device can identify a port partner that sources a Fixed power profile
as defined by the USB PD specification.

Test: run cts-v-host -m CtsUsbTypecTestCases --module-arg CtsUsbTypecTestCases:instrumentation-arg:ENABLE_MANUAL:=true
Bug: 450113925
Flag: android.hardware.usb.flags.enable_power_profile_reporting
Change-Id: Ie945b3ce0ab755e33b431887dc40f9ec98c61e79
Signed-off-by: RD Babiera <rdbabiera@google.com>