Bugfix: Move PersistableBundle byte array test behind flag.
In ag/36835652, getByteArray/putByteArray are introduced behind a flag.
In the same topic, in ag/36835653 the test is unconditional of the flag value.
When the flag is not enabled, the test fails.

This separates the new behavior in a flag-gated test.

Bug: 449713775
Bug: 437280042
Bug: 462031489
Flag: android.app.privatecompute.flags.enable_pcc_framework_support
Change-Id: Ia06b178b72fd301eef64d1b6ed329124b95acf94