Improve touch event verification in WindowUntrustedTouchTesttouch
Replaces the `CountDownLatch` in `WindowUntrustedTouchTestBase` with a
`BlockingQueue`-based event verifier for more robust touch event
testing.

The new verifier provides a better waiting and blocking mechanism and
allows for more detailed event inspection. This was not possible with
the previous `CountDownLatch` implementation, which could only signal
the reception of an event but not store the event itself.

Test: atest CtsWindowManagerDeviceInput:WindowUntrustedTouchTest
Test: atest --user-type secondary_user CtsWindowManagerDeviceInput:WindowUntrustedTouchTest
Test: atest --user-type secondary_user_on_secondary_display CtsWindowManagerDeviceInput:WindowUntrustedTouchTest
Test: atest CtsWindowManagerSdk35TestCases
Test: presubmit
Flag: TEST_ONLY
Bug: 412335432
Change-Id: I879f26e65bfdcb5c9c51509b27fc033a1ee9fbc3