Fix test given new PersistableBundle capability
In ag/36835652, PersistableBundle gets the ability to get/put byte
arrays. In this test, reading the `invalid` byte array will now succeed
(nothing is invalid except the name).

Looking at the CVE and b/247513680, this seems fine; the root cause of
the CVE was a crash when parsing the bundle containing a (then
unsupported) byte array. This is now supported and does not cause an
exception.

Bug: 449713775
Bug: 437280042
Bug: 462298510
Flag: android.app.privatecompute.flags.enable_pcc_framework_support
Change-Id: I64d4c6106bdffe7b9d737efeb9f000935c9b1c1f