mediav2: fix test for CDD 5.1.7/C-2-1
As per CDD Sec 5.1.7/C-2-1, if device implementation advertise HDR
profile support via Display.HdrCapabilities they MUST have a decoder
capable of handling that profile. The test is verifying the logical
converse of this requirement. The test is updated to be more in line
with the CDD in 26q2. The test is removed from older revisions.

Bug: 384649922
Test: atest android.mediav2.cts.CodecInfoTest
Flag: TEST_ONLY
Change-Id: Ib86630ba18202ceaf40187a4e9123fa95d5351b6
Merged-In: Ib86630ba18202ceaf40187a4e9123fa95d5351b6