mediav2: fix test for CDD 5.1.7/C-2-1
As per CDD Sec 5.1.7/C-2-1, if device implementation advertise HDR
profile support via Display.HdrCapabilities they MUST have a decoder
capable of handling that profile. The test is verifying the logical
converse of this requirement. The test is updated to be more in line
with the CDD.

Bug: 435229083
Test: atest android.mediav2.cts.CodecInfoTest
Test: atest android.mediav2.cts.CodecListTest
Flag: TEST_ONLY
Change-Id: I0205ce8c9a6b2307144e22c82c04c6cb99ceb0a4