Switch back to currentUser in MultiUser IME tests
A few MultiUser IME tests require switching to a different full user.
However, at the end of the test, this remains the current user on the
device. This adds a switch back to the initial user after the test
finishes, to avoid test isolation issues when later trying to start an
activity in the profile of a non-current full user.

Flag: EXEMPT TESTFIX
Bug: 461640579
Test: atest MultiUserMockImeTest#testFullUserSwitching
  MultiUserMockImeTest#testProfileSwitchingThenFullUserSwitching
  MultiUserTest#testAdditionalUser
Change-Id: I62b570eb78431d273c2388ea3e468ca8be4f266d