Fix(Telephony): Skip voicemail CTS tests on watches
The TelephonyManagerReadPhoneStatePermissionTest fails on watch devices
because they do not support voicemail APIs, leading to a
NullPointerException when the test attempts to access them.

This change adds a check to determine if the test is running on a watch
device and skips the voicemail-related API calls if it is.

Flag: EXEMPT
Bug: 458254038
Test: TelephonyManagerReadPhoneStatePermissionTest
Change-Id: Ideb9b7558439a31a7f238c93f99a0b25736a3708