Restrict AppcompatState tests from running on all foldable device states
The appcompat state tests modify the device state of the device
which causes failures and infra issues if paired with being ran
on all device states.

This change pulls these tests into it's own separate module
with a specific configuration to not run through the device states.

Flag: EXEMPT TEST_ONLY
Bug: 459920811
Test: atest CtsAppcompatStateAtomHostTestCases
Change-Id: Ie4aa31c325ea0828e5481ccdf9ffd620a91c35dd