ADMIN_USER tests use @RequireRunOnInitialUser
UserTest has four tests that are expecting the test user to be an ADMIN.
Two of these tests make sure to RequireRunOnInitialUser, which is
currently indeed an Admin user.
Two of these tests, however, have instead been using
RequireRunOnSystemUser, which is only okay if the System user is an
Admin. That assumption is now broken, so we need to switch these to the
safer RequireRunOnInitialUser that the other tests use.

Really, all four of these tests should use @RequireRunOnAdminUser, but
that doesn't exist at present; we add a TODO.

Fixes: 461626957
Bug: 461900451
Test: atest CtsDevicePolicyTestCases:android.devicepolicy.cts.UserTest#removeUser_disallowRemoveUserIsNotSet_isRemoved,removeUser_disallowRemoveUserIsSetOnAdminUser_returnsFalse
Flag: EXEMPT BUGFIX
Change-Id: Ia0bea8ed947db84a833e115d976e2865cfe41d7b