Explicitly resolve a web URL Intent to app
This CL changes the `launchActivity` method to explicitly launch an
activity in an app rather than a web browser with the Intent with a web
URL.
On some web-preferred devices (e.g., desktop), Intent with a web link
(and ACTION_VIEW) triggers a popup to navigate a user to a browser from
the app, which makes the CTS test to fail.
`FLAG_ACTIVITY_REQUIRE_NON_BROWSER` indicates that an Intent sender
expects the web URL Intent is resolved to an app rather than a browser.
So the app-to-web popup won’t be shown and the CTS test can pass.

Fix: 461367323
Flag: EXEMPT TEST_ONLY
Test: atest CtsInputMethodServiceHostTestCases
Change-Id: Ifbd9960a52ab9dfa5e778f34d1669e97a3939f8c