Fix flakiness in some OutputSwitcherTests
The getUniqueId helper was assuming that we'd always get an
onRoutesUpdated callback where we can find the assigned unique id for
the route, but it seems every once and a while the event had already
been sent by the time we register our callback, so the test times out.

This CL fixes the problem by switching to just polling the getRoutes()
results.

Bug: 465412486
Flag: TEST_ONLY
Test: atest --rerun-until-failure 50 OutputSwitcherTest
Change-Id: I2e118bd91cc0bea9fdf817029e8466c42cf27b1a