Add CTS tests for Native App Zygote
This commit implements the integration test for native services using
Native App Zygote (go/native-app-zygote). As most of the logic overlaps
between the existing native service tests and the new Native App Zygote
tests, they are implemented as child classes of NativeServiceTestBase.

We test two things; 1) the specified shared library is preloaded and the
specified function is called, and 2) multiple services with
`useAppZygote="true"` are spawned from the same Native App Zygote
process.

Test: atest CtsNativeServiceTest
Bug: 461341143
Change-Id: Iad4b22a82772709241e0b14f201ce4c54dfbb919