Introduces NativeServiceGTestRunner
To test the behavior of NDK APIs, we run the corresponding CTS tests in
native services. Since such tests need to be implemented in native
languages, we assume they are implemented as GTests. This commit
introduces NativeServiceGTestRunner, which runs a native service that
calls `RUN_ALL_TESTS` to execute all the GTest test cases. The user of
this class should link their own GTests against the libraries defined
here.

Bug: 461305201
Test: atest CtsTextNativeTestCases
Change-Id: Id6d1afc65f082945708c4e0f7d62a06ad97751aa