Update CTS tests for fullscreen disconnect.
With the addition of FullscreenDisconnectHandler, tasks are now reparented to the back on disconnect instead of the front. This means many tests are outdated as they assume tasks will be reparented to the front and dispatch a config change. This CL skips or updates tests where applicable when the fullscreen disconnect flag is enabled to adhere to the new assumption that tasks will be moved to the back.

Bug: 434035592
Test: atest with and without fullscreen disconnect flag
Flag: com.android.window.flags.enable_display_disconnect_fullscreen
Change-Id: I99c3fafedaf58b95f8a1c18918eab326fc215b01