Allow testSetApplicationNightMode run during daytime
This CL explicitly specifies the system night mode before executing the
applicationNightMode test.
Currently, the test runs against the default system night mode which
varies on the device configuration. If the default is set to
MODE_NIGHT_AUTO, the actual resolved night mode can fluctuate based on
the time of day. However, the test incorrectly assumes the night mode is
always enabled.
The purpose of this test is to verify the behavior of
applicationNightMode so we should test the behavior of the cases where
the system night mode is enabled and the application night mode is
disabled, and vice versa.

Flag: EXEMPT TEST_ONLY
Bug: 467477343
Bug: 383590347
Test: SplashscreenTests#testSetApplicationNightMode_*
Change-Id: Ifc49233449b5d3ddbc3438fa688019382c33d15b