mediav2: correct support requirements in codec encoder test
The intent of CodecEncoderTest is to verify all encoder components for
various media codec api. The test is picking low resolution encoding
configurations during testing with the idea that components on slow-cpu
devices, components with very limited capabilities will also be covered.
However a resolution that works for all components seems to be not
possible.

The best way to address this is pick configuration at run time based on
the advertised capabilities. This is handled in android 26q2. Legacy CTS
tests may skip if configuration is not supported.

Bug: 465641721
Bug: 461706007
Test: atest CtsMediaV2TestCases:CodecEncoderTest
Flag: TEST_ONLY
Change-Id: Iab921a40463b7af2de90ac17a2317f8f5027586e
Merged-In: Iab921a40463b7af2de90ac17a2317f8f5027586e