Ensure missing flag does not cause test crash.
Due to the unfortunate new release process for the CTS suite, we can end up
in a situation where flags are being used on devices where they are not
declared.  We have historically just marked those tests as known failures
to skip.  Instead, I have defined a new "isFlagEnabled" function that
will capture the NoSuchMethodError exception and return `false` instead.
Since the flag doesn't exist it is reasonable to assume that it will be
false if not present.

Test: Fixes a test failure.
Bug: 467858714
Flag: EXEMPTY BUG_FIX
Change-Id: Ie3b9abe3d7ae9a00dce157cc80c85b8e4512c7d2