[cts] fix CtsPackageInstallerCUJUpdateSelfTestCases after new updating screen
The newly introduced "updating..." screen is in the same task as the
app. Therefore we should start the installer dialog in a new task to
prevent it from being dismissed by the new screen. Further
discussions are needed to determine the optimal UX flow for
self-updating apps.

Also refactors the test util methods to use a Builder for readability.

BUG: 470062695
FIXES: 470062695
Test: atest
android.packageinstaller.criticaluserjourney.cts.UpdateSelfViaIntentTest#noLaunchGrantPermission_update_success
FLAG: TEST_ONLY

Change-Id: I2a1a5293611b425ec9af40e2c2913e86f4bbec22