Fix testDataSaverMode_blacklisted for AAOS
AAOS seems to have a slightly different app lifecycle, causing the
broadcast receivers to get leaked, which is causing CTS failures.
Unregister the receiver in onDestroy.

This does not completely fix the test, but it seems the remaining errors
are something in AAOS sysui, and the bug has been handed over to that
team.

And fix ACTIVTIY spelling.

Bug: 457827865
Test: atest com.android.cts.netpolicy.HostsideRestrictBackgroundNetworkTests#testDataSaverMode_blacklisted
Change-Id: I94bdc09ac06b673da3937d3ad15893a249925f24