Move GPU descriptors extraction into Utils, and add tests.
Improve modularity and test coverage of the CTS test, and add more details to some of the errors.
This change also makes the test faster as it reduces the number of times we scan the trace.

I originally planned to include these in the GPU utilisation counter change, but it grew too large so splitting out for easier review.

Test: atest hostsidetests/graphics/gpuprofiling [A]
Flag: EXEMPT TEST_ONLY
Bug: 448868135
Change-Id: Ia951673f659d156f321974beac59c5c29ea92aee