Make sure the app is stopped before it's launched.
With the recent package install flow, if the app is visible at the moment of package install then the app will stay visible and be restarted. To make sure that the tests don't hang try to launch the app again, add "-S" option which will stop the activity before launching it.

Bug: 464502389
Test: atest android.incrementalinstall.cts.IncrementalInstallTest
Flag: EXEMPT TEST_ONLY
Change-Id: I68649f7c9f9c8b523d809d1ca9f40011cee43b9e