Ensure exception thrown during contact insertion is logged.
We catch this to prevent the test runner crashing, but it masks the root
cause of the exception.  I saw this come up in b/469343145 and some
extra digging showed it was a permissions denial in the provider. Logging
the exception would maek that more clear though.

Test: Run CTS tests.
Flag: EXEMPT TEST_ONLY
Fixes: 469749179
Change-Id: I3676f2c6775b85fdf17e9210cca8712a27b2b38a