Modify the multi client tests to use reflection
Camera2 multi-client feature was introduced in Android 25Q2.
This feature is backported to Android 15 Automotive Car branch to
enable automotive camera use cases needing multiple clients to access
the same camera. The CTS tests for this feature calls new camera apis
which might not be available on devices with android 15 build which does
not have this feature backported. Hence modify the CTS tests to use java
reflection to only call the new apis when present in the code on devices
which has this feature backported. Also skipping the tests if device is
not automotive.

Bug: 444512779
Test: Run CTS tests
android.hardware.multiprocess.camera.cts.SharedCameraTest and
android.hardware.multiprocess.camera.cts.SharedSessionConfigurationTest
on android 15 build which includes the backported multi-client feature
and on the builds which does not have this feature

Change-Id: I0855c095d2433e437500a4eb7e1643a84067b860