Fix testCreateMirror test.
The testCreateMirror test launches a fullscreen activity, attaches the
mirror of the activity to a virtual display, and then verify the content
presented on the virtual display.

When the mirror size is smaller than the virtual display size, there
will be empty black areas on the virtual display, which break the test.

This change fixes that issues by limiting the checking bounds to the
minimum of mirror size and the virtual display size.

Test: locally ran the test with acloud.
Bug: 470176229
Flag: EXEMPT BUGFIX.

Change-Id: I39897da4f145b210c72e18dbcb36053ac726391b