WebView: remove documentation references to SDK tests
No change to logic, only documentation and build file for tests. This
updates documentation to remove some references to the SDK runtime test
suite (which has been deleted already). I'm leaving the bulk of the
documentation for the helper library only because the library still
exists and we may need this documentation to debug in the meantime until
the library can be fully deleted. This guide may also be helpful for
refactoring, since we can follow the steps in reverse to migrate off of
the helper library.

Bug: 453666068
Test: N/A
Flag: TEST_ONLY
Change-Id: I3cfdd209def79bbb252315ece947191a5fa151f7