Wait for animations to settle before checking keyguard
Currently, the LockScreenSession close method relies on a sleep-wakeup
cycle for removing keyguard, and then performs an unlockDevice if the
keyguard is still up. However, before checking whether this is needed or
not, the method does not wait for the animations resulting from
sleep-wakeup cycle to stabilize. This change adds this wait to ensure
that the unlockDevice steps are only performed if necessary.

On Android XR, this resulted in solving a race condition. Before this
change, the test sometimes erroneously tapped on SysUIVisibilityBarrier
because the animations were still in progress, which resulted in it
being put to the top of task stack. As a result, the test activity was
minimized, causing assertion failure when the test checked for it to be
in ON_RESUME state. After this change, the keyguard is always removed
due to the sleep-wakeup cycle, resulting in the test never needing to
call the offending method midway through a running transition.

Bug: 447452305
Test: atest android.server.wm.lifecycle.ActivityLifecycleKeyguardTests
Change-Id: Ie7c518bbe66393c28f2f960597326a661eebe560