ADMIN_USER tests use @RequireRunOnInitialUser only if flag enabled
In Ia0bea8ed947db84a833e115d976e2865cfe41d7b we updated these tests so
that on HSUM they would work when user 0 is not an Admin. However, user
0 being a non-Admin only applies right now if the hsu_not_admin flag is
enabled. So when that flag is disabled, the updated tests fail: user 0
is an Admin and gets the restriction, but the test now runs on user 10
instead and therefore fails. We fix this by making flag-enabled and
flag-disabled versions of this test.

Fixes: 467802613
Bug: 461626957
Flag: android.multiuser.hsu_not_admin
Test: atest CtsDevicePolicyTestCases:android.devicepolicy.cts.UserTest#removeUser_disallowRemoveUserIsNotSet_isRemoved,removeUser_disallowRemoveUserIsSetOnAdminUser_returnsFalse
Change-Id: Icb19795ea665d3586434d4466de5a12d9d1dfe12