Remove axis events from Razer key tests
The key tests should only be checking the key-related functionality
(even though in practice, the device would produce both simultaneously).
Prior to this CL, the test was injecting key event as well as an axis
event - which would cause the test to fail occasionally.

The failure was not permanent, because the gamepad events were batched -
and the consumption would only happen at the vsync boundaries.
Therefore, the test would only fail occasionally on slow devices because
keys were processed slower, and would sometimes come in after the vsync.

Bug: 470791228
Test: atest android.hardware.input.cts.tests.RazerJunglecatBluetoothTest#testAllKeys
Flag: EXEMPT TEST_ONLY
Change-Id: Ic9d752ee366f8431b78ce2dac5354746cfbed0a7