Wait before checking that there weren't extraneous events
In the previous version of the code, the check 'assertNoMoreEvents' was
instantaneous. This means that it would only pass if there weren't any
events at that moment. However, when the events are batched (such as
joystick MOVE events), the events might not be consumed until the next
vsync. This caused the test to flake occasionally on slow devices.

In this CL, we add a small, 100ms wait to this check. This ensures that
the events have plenty of time to arrive, and makes the test more
predictable.

Bug: 470791228
Test: atest CtsHardwareTestCases
Flag: EXEMPT TEST_ONLY
Change-Id: Ib69b7276f9a5130af5586f243650c207dd737f55