Add CTS test for PackageManager#GET_APP_LOCK_INFO
Providing the GET_APP_LOCK_INFO flag in
PackageManager#getApplicationInfo returns an ApplicationInfo object
where isAppLockSupported is set to false for headless apps and
isAppLockEnabled is set to true only if the package is App Lock enabled
and a screen lock is configured on the device. If no screen lock is
present, isAppLockEnabled is set to false.

Bug: 471705717
Test: atest ApplicationInfoTest#testGetApplicationInfo_withAppLockEnabled_returnsIsAppLockEnabledTrue
Test: atest ApplicationInfoTest#testGetApplicationInfo_withoutScreenLock_returnsIsAppLockEnabledFalse
Test: atest ApplicationInfoTest#testGetApplicationInfo_withHeadlessApp_returnsIsAppLockSupportedFalse
Test: atest ApplicationInfoTest#testGetApplicationInfo_exemptPackage_returnsIsAppLockSupportedFalse
Test: atest ApplicationInfoTest#testGetApplicationInfo_supervisedUser_returnsIsAppLockSupportedFalse
Test: atest ApplicationInfoTest#testGetApplicationInfo_managedProfile_returnsIsAppLockSupportedFalse
Test: I74100010507745813
Flag: EXEMPT TEST_ONLY
Change-Id: Ide9aec285583edb03d7261b99fae783402071c32