Prime testGetAllCellInfo before testing
Reduce the flake rate of testGetAllCellInfo by priming the
test with an updated version of the cell info from the cache.
This addresses an issue where the CellInfo cache on the AP has
not already been updated after the UE reselected to the current
cell. On devices that proactively send an update from the modem,
this won't be an issue; but, it can still happen on devices
where for some reason the cell info was just updated prior to
the test, forcing the cache to return an old value, but it should
be extremely rare due to devices undergoing CTS not moving.

Bug: 467797099
Test: atest TelephonyManagerTest#testGetAllCellInfo
Flag: EXEMPT TEST-ONLY
Change-Id: I4a908486a66c5ade428dbe2e41b05d25101426e8